Moving Adapter, Converter and Matcher to separate package
