+++ src/main/java/org/cqfn/astranaut/core/algorithms/conversion/package-info.java
@@ -21,21 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cqfn.astranaut.core.base;
 
 /**
- * Interface for converters that check one rule described in DSL
- * and convert the initial AST built to the specified target format.
- *
+ * This package contains structures and algorithms for converting a syntax tree\
+ *  to some other syntax tree.
  * @since 1.0
  */
-public interface Converter {
-    /**
-     * Converts an initial AST to the target format.
-     *
-     * @param node The root of the AST to be converted
-     * @param factory The node factory
-     * @return A new node
-     */
-    Node convert(Node node, Factory factory);
-}
+package org.cqfn.astranaut.core.algorithms.conversion;
