+++ src/test/java/org/cqfn/astranaut/core/example/javascript/rules/Matcher1.java
@@ -25,7 +25,7 @@ package org.cqfn.astranaut.core.example.javascript.rules;
 
 import java.util.List;
 import java.util.Map;
-import org.cqfn.astranaut.core.base.Matcher;
+import org.cqfn.astranaut.core.algorithms.conversion.Matcher;
 import org.cqfn.astranaut.core.base.Node;
 
 /**
