+++ src/test/java/org/cqfn/astranaut/core/example/javascript/rules/Rule0.java
@@ -26,8 +26,8 @@ package org.cqfn.astranaut.core.example.javascript.rules;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
+import org.cqfn.astranaut.core.algorithms.conversion.Converter;
 import org.cqfn.astranaut.core.base.Builder;
-import org.cqfn.astranaut.core.base.Converter;
 import org.cqfn.astranaut.core.base.DummyNode;
 import org.cqfn.astranaut.core.base.Factory;
 import org.cqfn.astranaut.core.base.Node;
